==> 1-load_data.ipynb <==
import numpy as np
from ucimlrepo import fetch_ucirepo 
import os
from sklearn.preprocessing import LabelEncoder
import pandas as pd

""""

Car evaluation
Breast cancer
Australian
Heart
Adult
Student Performance
banknote
sonar
SPAM
WINE

"""

namelist = ["car","breast",
            "australian","heart","adult","student",
            "banknote","sonar","spam","wine"]
#namelist = ["adult"]
for dataname in namelist:
    
    datapath = f'dataset/{dataname}'
    os.makedirs(datapath, exist_ok=True)

    load_data(dataname)

# wine
def load_data(dataname):
    if dataname == "wine":
        data = fetch_ucirepo(id=186) 
        # Split the data into features and label
        shuffled_df = data.data.original.sample(frac=1).reset_index(drop=True)
        features = shuffled_df.iloc[:, :-1]
        label = shuffled_df.iloc[:, -1]
        label_encoder = LabelEncoder()
        label = label_encoder.fit_transform(label)

    elif dataname == "spam":
        data = fetch_ucirepo(id=94) 
        shuffled_df = data.data.original.sample(frac=1).reset_index(drop=True)
        features = shuffled_df.iloc[:, :-1]
        label = shuffled_df.iloc[:, -1]
    
    elif dataname == "sonar":
        data = fetch_ucirepo(id=151) 
        shuffled_df = data.data.original.sample(frac=1).reset_index(drop=True)
        features = shuffled_df.iloc[:, :-1]
        label = shuffled_df.iloc[:, -1]
        label_encoder = LabelEncoder()
        label = label_encoder.fit_transform(label)

    elif dataname == "banknote":
        data = fetch_ucirepo(id=267) 
        shuffled_df = data.data.original.sample(frac=1).reset_index(drop=True)
        features = shuffled_df.iloc[:, :-1]
        label = shuffled_df.iloc[:, -1]

    elif dataname == "australian":
        data = fetch_ucirepo(id=143)
        shuffled_df = data.data.original.sample(frac=1).reset_index(drop=True)
        features = shuffled_df.iloc[:, :-1]
        label = shuffled_df.iloc[:, -1]

        
    elif dataname == "heart":
        data = fetch_ucirepo(id=45)
        shuffled_df = data.data.original.sample(frac=1).reset_index(drop=True)
        features = shuffled_df.iloc[:, :-1]
        label = shuffled_df.iloc[:, -1]


    elif dataname == "adult":
        data = fetch_ucirepo(id=2)
        shuffled_df = data.data.original.sample(frac=1).reset_index(drop=True)
        features = shuffled_df.iloc[:, :-1]
        label = shuffled_df.iloc[:, -1]
        label_encoder = LabelEncoder()
        label = label_encoder.fit_transform(label)


    elif dataname == "student":
        data = fetch_ucirepo(id=320)
        shuffled_df = data.data.original.sample(frac=1).reset_index(drop=True)
        shuffled_df['G3_category'] = pd.cut(shuffled_df['G3'], bins=5,labels=[0, 1, 2, 3, 4,])
        features = shuffled_df.iloc[:, :-4]
        label = shuffled_df['G3_category'] 
        
    elif dataname == "car":
        data = fetch_ucirepo(id=19)
        shuffled_df = data.data.original.sample(frac=1).reset_index(drop=True)
        features = shuffled_df.iloc[:, :-1]
        label = shuffled_df.iloc[:, -1]
        label_encoder = LabelEncoder()
        label = label_encoder.fit_transform(label)

    elif dataname == "breast":
        data = fetch_ucirepo(id=14)
        shuffled_df = data.data.original.sample(frac=1).reset_index(drop=True)
        features = shuffled_df.iloc[:, :-1]
        label = shuffled_df.iloc[:, -1]
        label_encoder = LabelEncoder()
        label = label_encoder.fit_transform(label)
    print(dataname,len(set(label)))
    #save_data(features,label,datapath)

def save_data(features,label,datapath):

    pd.DataFrame(features).to_csv(f'{datapath}/features.csv', index=False)
    pd.DataFrame(label).to_csv(f'{datapath}/label.csv', index=False)

    print(f"{datapath} Save CSV")
    # Convert to NumPy arrays
    features_np = features
    label_np = label


    try:
        print(datapath)
        np.save(f'{datapath}/features.npy', features_np)
        np.save(f'{datapath}/label.npy', label_np)
        print("Files saved successfully")
    except Exception as e:
        print("Fail to save npy")
        print(f"Error: {e}")
    

## deal with missing, Need to delete later

import numpy as np
import pandas as pd

namelist = ["breast", "heart", "adult"]
namelist = ["adult"]
def count_and_replace_nans(data):
    has_nan = False
    total_nan_count = 0
    column_nan_counts = np.zeros(data.shape[1], dtype=int)
    
    for i in range(data.shape[0]):
        for j in range(data.shape[1]):
            try:
                if np.isnan(data[i, j]):
                    has_nan = True
                    total_nan_count += 1
                    column_nan_counts[j] += 1
                    # Check if column is categorical (string type)
                    if isinstance(data[0, j], str) or isinstance(data[1, j], str):
                        data[i, j] = 'Na'
                    else:
                        data[i, j] = 0  # Replace numerical NaN with 0
            except TypeError:
                continue
    return has_nan, total_nan_count, column_nan_counts, data

for dataname in namelist:
    datapath = f'dataset/{dataname}'
    data = np.load(f'{datapath}/features.npy', allow_pickle=True)
    data1 = pd.read_csv(f'{datapath}/features.csv')


    # Check if there are any NaN values and count them, replace as needed
    has_nan, total_nan_count, column_nan_counts, updated_data = count_and_replace_nans(data)
    
    print(f"Contains NaN: {has_nan}")
    print(f"Total number of NaN values: {total_nan_count}")
    print(f"Number of NaN values in each column: {column_nan_counts}")

    # Save the updated data back to .npy file
    np.save(f'{datapath}/features.npy', updated_data)

    print('Data saved')

["car","breast",
            "australian","heart","adult","student",
            "banknote","sonar","spam","wine"]

## Generate Column Type

import json

namelist = ["banknote","sonar","spam","wine"]
for dataname in namelist:
    info = {}
    datapath = f'dataset/{dataname}'
    data1 = pd.read_csv(f'{datapath}/features.csv')
    for i,j in enumerate(data1.columns):
     info[i] = "numerical"

         # Save the info dictionary to a JSON file
    json_path = f'{datapath}/column_info.json'
    with open(json_path, 'w') as json_file:
        json.dump(info, json_file, indent=4)


car_mapping = {
            "buying": {"vhigh": 4, "high": 3, "med": 2, "low": 1},
            "maint": {"vhigh": 4, "high": 3, "med": 2, "low": 1},
            "doors": {"2": 2, "3": 3, "4": 4, "5more": 5},
            "persons": {"2": 2, "4": 4, "more": 5},
            "lug_boot": {"small": 1, "med": 2, "big": 3},
            "safety": {"low": 1, "med": 2, "high": 3}
        }
breast_mapping = {
    "age": {'50-59': 3, '40-49': 2, '60-69': 4, '30-39': 1, '70-79': 5, '20-29': 0},
    "menopause": {'premeno': 0, 'ge40': 1, 'lt40': 2},
    "tumor-size": {'30-34': 6, '25-29': 5, '20-24': 4, '15-19': 3, '10-14': 2, '40-44': 8, '35-39': 7, '50-54': 10, '0-4': 0, '5-9': 1, '45-49': 9},
    "inv-nodes": {'0-2': 0, '3-5': 1, '6-8': 2, '9-11': 3, '12-14': 4, '15-17': 5, '24-26': 6},
    "node-caps": {'no': 0, 'yes': 1},
    "deg-malig": {2: 0, 3: 1, 1: 2},
    "breast": {'left': 0, 'right': 1},
    "breast-quad": {'left_low': 0, 'left_up': 1, 'right_up': 2, 'right_low': 3, 'central': 4},
    "irradiat": {'no': 0, 'yes': 1}
}
australian_mapping = {
    "A1":{0:0,1:1},
    "A4":{1:0, 2:1, 3:2},
    "A5":{1:0,2:1,3:2,4:3,5:4,6:5,7:6,8:7,9:8,10:9,11:10,12:11,13:12,14:13},
    "A6":{1:0,2:1,3:2,4:3,5:4,6:5,7:6,8:7,9:8},
    "A8":{0:0,1:1},
    "A9":{0:0,1:1},
    "A11":{0:0,1:1},
    "A12":{1:0, 2:1, 3:2 }
}

heart_mapping = {
"sex":{1: 1, 0: 0},
"cp":{4: 4, 3: 3, 2: 2, 1: 1},
"fbs":{0: 0, 1: 1},
"restecg":{0: 0, 2: 2, 1: 1},
"exang":{0: 0, 1: 1},
"slope":{1: 1, 2: 2, 3: 3},
"ca":{0.0: 0, 1.0: 1, 2.0: 2, 3.0: 3},
"thal":{3.0: 3, 7.0: 7, 6.0: 6},       
}

adult_mapping = {
   #nominal
"workclass": {'Private': 1, 'Self-emp-not-inc': 2, 'Local-gov': 3, 'State-gov': 4, '?': 5, 'Self-emp-inc': 6, 'Federal-gov': 7, 'Without-pay': 8, 'Never-worked': 0},
# ordinal
"education": {"Preschool": 0,
              "1st-4th": 1,
              "5th-6th": 2,
               "7th-8th": 3,
               "9th": 4,
               "10th": 5,
               "11th": 6,
               "12th": 7,
               "HS-grad": 8,
               "Some-college": 9,
               "Assoc-voc": 10,
               "Assoc-acdm": 11,
               "Bachelors": 12,
               "Masters": 13,
               "Prof-school": 14,
               "Doctorate": 15},
# nominal
"marital-status": {'Married-civ-spouse': 1,
                   'Never-married': 2,
                   'Divorced': 3,
                   'Separated': 4,
                   'Widowed': 5,
                   'Married-spouse-absent': 6,
                   'Married-AF-spouse': 0},
# nominal
"occupation": {'Prof-specialty': 1, 
               'Craft-repair': 2, 
               'Exec-managerial': 3, 
               'Adm-clerical': 4, 
               'Sales': 5, 
               'Other-service': 6, 
               'Machine-op-inspct': 7, 
               'Transport-moving': 8, 
               'Handlers-cleaners': 9, 
               '?': 10, 
               'Farming-fishing': 11, 
               'Tech-support': 12, 
               'Protective-serv': 13, 
               'Priv-house-serv': 14, 
               'Armed-Forces': 0},
# nominal
"relationship": {'Husband': 1,
                  'Not-in-family': 2,
                  'Own-child': 3,
                  'Unmarried': 4, 
                  'Wife': 5, 
                  'Other-relative': 0},
# nominal
"race": {'White': 1, 'Black': 2, 
         'Asian-Pac-Islander': 3, 'Amer-Indian-Eskimo': 4, 
         'Other': 0},
# nominal
"sex": {"Female":0, 
        "Male" : 1},
# nominal
"native-country": {'United-States': 1, 'Mexico': 2, '?': 3, 'Philippines': 4, 'Germany': 5, 
                   'Puerto-Rico': 6, 'Canada': 7, 'El-Salvador': 8, 'India': 9, 'Cuba': 10, 
                   'England': 11, 'China': 12, 'South': 13, 'Jamaica': 14, 'Italy': 15, 
                   'Dominican-Republic': 16, 'Japan': 17, 'Guatemala': 18, 'Poland': 19, 'Vietnam': 20, 
                   'Columbia': 21, 'Haiti': 22, 'Portugal': 23, 'Taiwan': 24, 'Iran': 25, 'Nicaragua': 26,
                     'Greece': 27, 'Peru': 28, 'Ecuador': 29, 'France': 30, 'Ireland': 31, 'Hong': 32, 
                     'Thailand': 33, 'Cambodia': 34, 'Trinadad&Tobago': 35, 'Outlying-US(Guam-USVI-etc)': 36, 
                     'Laos': 37, 'Yugoslavia': 38, 'Scotland': 39, 'Honduras': 40, 'Hungary': 41, 'Holand-Netherlands': 0}
}

student_mapping = {
   "school": {'GP': 0, 'MS': 1},
   "sex":{'F': 0, 'M': 1},
   "address": {'U': 0, 'R': 1},
   "famsize":{'GT3': 0, 'LE3': 1},
   "Pstatus":{'T': 0, 'A': 1},
   "Medu": {2: 2, 4: 4, 1: 1, 3: 3, 0: 0}, # ordinal
   "Fedu": {2: 2, 1: 1, 3: 3, 4: 4, 0: 0},# ordinal
   "Mjob":{'other': 4, 'services': 0, 'at_home': 1, 'teacher': 2, 'health': 3},
   "Fjob":{'other': 4, 'services': 0, 'at_home': 1, 'teacher': 2, 'health': 3},
   "reason":   {'course': 0, 'home': 1, 'reputation': 2, 'other': 3},
   "guardian":{'mother': 0, 'father': 1, 'other': 2},
   "traveltime": {1: 0, 2: 1, 3: 2, 4: 3},# ordinal
   "studytime":{2: 1, 1: 0, 3: 2, 4: 3},# ordinal
   "failures":{0: 0, 1: 1, 2: 2, 3: 3},# ordinal
   "schoolsup":{'no': 0, 'yes': 1},
   "famsup":{'no': 0, 'yes': 1},
   "paid":{'no': 0, 'yes': 1},
   "activities":{'no': 0, 'yes': 1},
   "nursery":{'no': 0, 'yes': 1},
   "higher":{'no': 0, 'yes': 1},
   "internet":{'no': 0, 'yes': 1},
   "romantic":{'no': 0, 'yes': 1},
   "famrel":{4: 3, 5: 4, 3: 2, 2: 1, 1: 0},# ordinal
   "freetime":{4: 3, 5: 4, 3: 2, 2: 1, 1: 0},# ordinal
   "goout":{4: 3, 5: 4, 3: 2, 2: 1, 1: 0},# ordinal
   "Dalc":{4: 3, 5: 4, 3: 2, 2: 1, 1: 0},# ordinal
   "Walc":{4: 3, 5: 4, 3: 2, 2: 1, 1: 0},# ordinal
   "health":{4: 3, 5: 4, 3: 2, 2: 1, 1: 0},# ordinal
}





namelist = ["car","breast",
            "australian","heart","adult","student"]
namelist = ["adult"]

for dataname in namelist:
   info = {}
   datapath = f'dataset/{dataname}'
   data1 = pd.read_csv(f'{datapath}/features.csv')
   if dataname == "car":
      for i,j in enumerate(data1.columns):
         info[i] = {"ordinal":car_mapping[j]}
         data1[j] = data1[j].map(car_mapping[j])
   elif dataname == "breast":
      for i,j in enumerate(data1.columns):
         if j in ["menopause" , "node-caps", "breast", "breast-quad", "irradiat" ]:
            info[i] = {"nominal":breast_mapping[j]}
         else:
            info[i] = {"ordinal":breast_mapping[j]}
         data1[j] = data1[j].map(breast_mapping[j])
            
   elif dataname == "australian":
      for i,j in enumerate(data1.columns):
         if j in australian_mapping.keys():
            info[i] = {"nominal":australian_mapping[j]}
            data1[j] = data1[j].map(australian_mapping[j])
         else:
            info[i] = "numerical"

   elif dataname == "heart":
      for i,j in enumerate(data1.columns):
         if j in heart_mapping.keys():
            info[i] = {"ordinal":heart_mapping[j]}
            data1[j] = data1[j].map(heart_mapping[j])
         else:
            info[i] = "numerical"


   elif dataname == "adult":
      for i,j in enumerate(data1.columns):
         if j in adult_mapping.keys():
            if j == "education":
               info[i] = {"ordinal":adult_mapping[j]}
            else:
               info[i] = {"nominal":adult_mapping[j]}
            data1[j] = data1[j].map(adult_mapping[j])
         else:
            info[i] = "numerical"
   elif dataname == "student":
      for i,j in enumerate(data1.columns):
         if j in student_mapping.keys():
            if j in [ "Medu","Fedu","traveltime" ,"studytime","failures","famrel"
                     "freetime","goout","Dalc","Walc","health"]:
               info[i] = {"ordinal":student_mapping[j]}
            else:
               info[i] = {"nominal":student_mapping[j]}
               data1[j] = data1[j].map(student_mapping[j])
         else:
            info[i] = "numerical"

   json_path = f'{datapath}/column_info.json'
   with open(json_path, 'w') as json_file:
      json.dump(info, json_file, indent=4)
    
   print(f"Column information for {dataname} saved to {json_path}")
    
   # Save the converted data to a new CSV file
   converted_data_path = f'{datapath}/features.csv'
   data1.to_csv(converted_data_path, index=False)
   print(f"Converted data for {dataname} saved to {converted_data_path}")
